factor tests
